Log wallet address and rig/worker ID
