+++ packages/web-app/src/modules/machine/MachineStore.ts
@@ -4,6 +4,12 @@ import { v4 as uuidv4 } from 'uuid'
 import * as Storage from '../../Storage'
 import { RootStore } from '../../Store'
 import { getPluginDefinitions } from '../salad-bowl/definitions'
+import {
+  Accounts,
+  BEAM_WALLET_ADDRESS,
+  ETH_WALLET_ADDRESS,
+  getNiceHashMiningAddress,
+} from '../salad-bowl/definitions/accounts'
 import { GpuInformation, MachineInfo } from './models'
 import { Machine } from './models/Machine'
 
@@ -90,7 +96,22 @@ export class MachineStore {
       return []
     }
 
-    const pluginDefinitions = getPluginDefinitions(machine, machineInfo)
+    // TODO: Use an existing copy of the plugin definition list generated by Salad Bowl.
+    const accounts: Accounts = {
+      ethermine: {
+        address: ETH_WALLET_ADDRESS,
+        workerId: machine.minerId,
+      },
+      flypoolBeam: {
+        address: BEAM_WALLET_ADDRESS,
+        workerId: machine.minerId,
+      },
+      nicehash: {
+        address: getNiceHashMiningAddress(machine.id),
+        rigId: machine.minerId,
+      },
+    }
+    const pluginDefinitions = getPluginDefinitions(accounts, machineInfo.platform ?? window.salad.platform)
     const gpus = machineInfo.graphics.controllers.map((gpu) => {
       const gpuMachineInfo: MachineInfo = {
         ...machineInfo,
@@ -129,7 +150,22 @@ export class MachineStore {
       return false
     }
 
-    const pluginDefinitions = getPluginDefinitions(machine, machineInfo)
+    // TODO: Use an existing copy of the plugin definition list generated by Salad Bowl.
+    const accounts: Accounts = {
+      ethermine: {
+        address: ETH_WALLET_ADDRESS,
+        workerId: machine.minerId,
+      },
+      flypoolBeam: {
+        address: BEAM_WALLET_ADDRESS,
+        workerId: machine.minerId,
+      },
+      nicehash: {
+        address: getNiceHashMiningAddress(machine.id),
+        rigId: machine.minerId,
+      },
+    }
+    const pluginDefinitions = getPluginDefinitions(accounts, machineInfo.platform ?? window.salad.platform)
 
     //Get all the CPU only plugin definitions
     const cpuPlugins = pluginDefinitions.filter((pluginDefinition) =>
