+++ packages/web-app/src/modules/salad-bowl/definitions/index.tsx
@@ -1,29 +1,12 @@
-import { Machine } from '../../machine/models/Machine'
-import { MachineInfo } from '../../machine/models/MachineInfo'
-import { Accounts, BEAM_WALLET_ADDRESS, ETH_WALLET_ADDRESS, getNiceHashMiningAddress } from './accounts'
+import type { Accounts } from './accounts'
 import { createLinuxPluginDefinitions } from './linux'
 import { createMacOSPluginDefinitions } from './macos'
-import { PluginDefinition } from './plugin-definitions'
+import type { PluginDefinition } from './plugin-definitions'
 import { createWindowsPluginDefinitions } from './windows'
 
-export const getPluginDefinitions = (machine: Machine, machineInfo: MachineInfo): PluginDefinition[] => {
-  const accounts: Accounts = {
-    ethermine: {
-      address: ETH_WALLET_ADDRESS,
-      workerId: machine.minerId,
-    },
-    flypoolBeam: {
-      address: BEAM_WALLET_ADDRESS,
-      workerId: machine.minerId,
-    },
-    nicehash: {
-      address: getNiceHashMiningAddress(machine.id),
-      rigId: machine.minerId,
-    },
-  }
-
+export const getPluginDefinitions = (accounts: Accounts, platform: string): PluginDefinition[] => {
   let pluginDefinitions: PluginDefinition[]
-  switch (machineInfo.platform ?? window.salad.platform) {
+  switch (platform) {
     case 'darwin':
       pluginDefinitions = createMacOSPluginDefinitions(accounts)
       break
